added greenhouse plants

+++ greenHousePlant.h
@@ -6,13 +6,13 @@
 #define GREENHOUSEPLANT
 class greenHousePlant{
 public:
-	std::vector<int> boost;
+	std::vector<float> boost;
 	std::string boostName;
 	std::string plantName;
 	float level;
 	float experience;
 
-	greenHousePlant(std::vector<int> pboost, std::string bName, std::string pName, float plevel, float pexperience)
+	greenHousePlant(std::vector<float> pboost, std::string bName, std::string pName, float plevel, float pexperience)
 	: boostName(bName), plantName(pName), level(plevel), experience(pexperience){
 		boost.assign(pboost.begin(), pboost.end());
 	}
